+++ vscode/src/commands/context/current-file.ts
@@ -3,6 +3,7 @@ import {
     ContextItemSource,
     TokenCounter,
     logError,
+    toRangeData,
     wrapInActiveSpan,
 } from '@sourcegraph/cody-shared'
 import * as vscode from 'vscode'
@@ -37,7 +38,7 @@ export async function getContextFileFromCurrentFile(): Promise<ContextItem[]> {
                     uri: document.uri,
                     content,
                     source: ContextItemSource.Editor,
-                    range: selection,
+                    range: toRangeData(selection),
                     size,
                 } satisfies ContextItem,
             ]