+++ vscode/src/commands/context/directory.ts
@@ -3,6 +3,7 @@ import {
     ContextItemSource,
     TokenCounter,
     logError,
+    toRangeData,
     wrapInActiveSpan,
 } from '@sourcegraph/cody-shared'
 import * as vscode from 'vscode'
@@ -64,7 +65,7 @@ export async function getContextFileFromDirectory(directory?: URI): Promise<Cont
                     uri: fileUri,
                     content,
                     source: ContextItemSource.Editor,
-                    range,
+                    range: toRangeData(range),
                     size,
                 })